+++ src/main/java/io/github/mmalykhin/hmsproxy/routing/RateLimitExceededException.java
@@ -1,0 +1,9 @@
+package io.github.mmalykhin.hmsproxy.routing;
+
+import org.apache.hadoop.hive.metastore.api.MetaException;
+
+final class RateLimitExceededException extends MetaException {
+  RateLimitExceededException(String message) {
+    super(message);
+  }
+}